factor out functions into providers and inject clients into functions. Note some nested dependency injection
